+++ event_routing_backends/processors/caliper/event_transformers/navigation_events.py
@@ -38,7 +38,6 @@ class NavigationEventsTransformers(CaliperTransformer):
         Returns:
             dict
         """
-        self.event_version = "1.0"
         self.backend_name = 'caliper'
         caliper_object = self.transformed_event['object']